functions/requirements.txt
# Core dependencies
firebase-admin>=6.2.0
openai>=1.0.0
python-dotenv>=1.0.0
firebase_functions~=0.1.0

# Cloud Functions dependencies
functions-framework>=3.4.0
google-cloud-secret-manager>=2.16.1
google-cloud-firestore>=2.11.0

# Utility packages
pydantic>=2.0.0
python-jose>=3.3.0
requests>=2.31.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0 